+++ database/migrations/2022_07_04_104832_add_indices_to_posts.php
@@ -1,0 +1,18 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    public function up()
+    {
+        Schema::table('posts', function (Blueprint $table) {
+            $table->index('title');
+            $table->index('slug');
+            $table->index('description');
+            $table->fullText('content');
+        });
+    }
+};