+++ src/Filament/Resources/FaqResource/Pages/CreateFaq.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\FaqResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use LaraZeus\Sky\Filament\Resources\FaqResource;
+
+class CreateFaq extends CreateRecord
+{
+    use CreateRecord\Concerns\Translatable;
+
+    protected static string $resource = FaqResource::class;
+}