+++ src/Filament/Resources/FaqResource/Pages/EditFaq.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\FaqResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use LaraZeus\Sky\Filament\Resources\FaqResource;
+
+class EditFaq extends EditRecord
+{
+    use EditRecord\Concerns\Translatable;
+
+    protected static string $resource = FaqResource::class;
+}