+++ src/Filament/Resources/FaqResource/Pages/ListFaqs.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\FaqResource\Pages;
+
+use Filament\Resources\Pages\ListRecords;
+use LaraZeus\Sky\Filament\Resources\FaqResource;
+
+class ListFaqs extends ListRecords
+{
+    use ListRecords\Concerns\Translatable;
+
+    protected static string $resource = FaqResource::class;
+}