+++ src/Filament/Resources/PageResource/Pages/CreatePost.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\PageResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use LaraZeus\Sky\Filament\Resources\PageResource;
+
+class CreatePost extends CreateRecord
+{
+    use CreateRecord\Concerns\Translatable;
+
+    protected static string $resource = PageResource::class;
+}