+++ src/Filament/Resources/TagResource/Pages/CreateTag.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\TagResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use LaraZeus\Sky\Filament\Resources\TagResource;
+
+class CreateTag extends CreateRecord
+{
+    use CreateRecord\Concerns\Translatable;
+
+    protected static string $resource = TagResource::class;
+}