+++ src/Filament/Resources/TagResource/Pages/EditTag.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\TagResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use LaraZeus\Sky\Filament\Resources\TagResource;
+
+class EditTag extends EditRecord
+{
+    use EditRecord\Concerns\Translatable;
+
+    protected static string $resource = TagResource::class;
+}